Move ConvertTo32 from prepostcleaning to utils.
