+++ xcp_d/interfaces/prepostcleaning.py
@@ -15,110 +15,9 @@ from nipype.interfaces.base import (
 
 from xcp_d.utils.confounds import _infer_dummy_scans, load_motion
 from xcp_d.utils.filemanip import fname_presuffix
-from xcp_d.utils.modified_data import _drop_dummy_scans, compute_fd, downcast_to_32
+from xcp_d.utils.modified_data import _drop_dummy_scans, compute_fd
 
 LOGGER = logging.getLogger("nipype.interface")
-
-
-class _ConvertTo32InputSpec(BaseInterfaceInputSpec):
-    bold_file = traits.Either(
-        None,
-        File(exists=True),
-        desc="BOLD file",
-        mandatory=False,
-        usedefault=True,
-    )
-    boldref = traits.Either(
-        None,
-        File(exists=True),
-        desc="BOLD reference file",
-        mandatory=False,
-        usedefault=True,
-    )
-    bold_mask = traits.Either(
-        None,
-        File(exists=True),
-        desc="BOLD mask file",
-        mandatory=False,
-        usedefault=True,
-    )
-    t1w = traits.Either(
-        None,
-        File(exists=True),
-        desc="T1-weighted anatomical file",
-        mandatory=False,
-        usedefault=True,
-    )
-    t1w_seg = traits.Either(
-        None,
-        File(exists=True),
-        desc="T1-space segmentation file",
-        mandatory=False,
-        usedefault=True,
-    )
-    t1w_mask = traits.Either(
-        None,
-        File(exists=True),
-        desc="T1-space mask file",
-        mandatory=False,
-        usedefault=True,
-    )
-
-
-class _ConvertTo32OutputSpec(TraitedSpec):
-    bold_file = traits.Either(
-        None,
-        File(exists=True),
-        desc="BOLD file",
-        mandatory=False,
-    )
-    boldref = traits.Either(
-        None,
-        File(exists=True),
-        desc="BOLD reference file",
-        mandatory=False,
-    )
-    bold_mask = traits.Either(
-        None,
-        File(exists=True),
-        desc="BOLD mask file",
-        mandatory=False,
-    )
-    t1w = traits.Either(
-        None,
-        File(exists=True),
-        desc="T1-weighted anatomical file",
-        mandatory=False,
-    )
-    t1w_seg = traits.Either(
-        None,
-        File(exists=True),
-        desc="T1-space segmentation file",
-        mandatory=False,
-    )
-    t1w_mask = traits.Either(
-        None,
-        File(exists=True),
-        desc="T1-space mask file",
-        mandatory=False,
-    )
-
-
-class ConvertTo32(SimpleInterface):
-    """Downcast files from >32-bit to 32-bit if necessary."""
-
-    input_spec = _ConvertTo32InputSpec
-    output_spec = _ConvertTo32OutputSpec
-
-    def _run_interface(self, runtime):
-        self._results["bold_file"] = downcast_to_32(self.inputs.bold_file)
-        self._results["boldref"] = downcast_to_32(self.inputs.boldref)
-        self._results["bold_mask"] = downcast_to_32(self.inputs.bold_mask)
-        self._results["t1w"] = downcast_to_32(self.inputs.t1w)
-        self._results["t1w_seg"] = downcast_to_32(self.inputs.t1w_seg)
-        self._results["t1w_mask"] = downcast_to_32(self.inputs.t1w_mask)
-
-        return runtime
 
 
 class _RemoveDummyVolumesInputSpec(BaseInterfaceInputSpec):
